backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
slowapi==0.1.8
python-multipart==0.0.9
newspaper3k>=0.2.8
lxml[html_clean]>=5.1.0